requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
requests==2.32.3
user-agents==2.2.0
cryptography==43.0.1
Werkzeug==3.0.3
Jinja2==3.1.4
click==8.1.7
itsdangerous==2.2.0
MarkupSafe==2.1.5
gunicorn==21.2.0
psycopg[binary]==3.2.1
